requirements.txt
pydantic==1.10.*
fastapi==0.110.*
uvicorn[standard]==0.29.*
prometheus_client==0.20.*
PyYAML==6.*
